♻️ TableEditor: extracted TableHead as file
also ActionsType made non default export on proptypes
