+++ lib/TableEditor/proptypes.js
@@ -12,4 +12,4 @@ const ActionsType = PropTypes.arrayOf(
     view: PropTypes.func,
   }),
 )
-export default ActionsType
+export { ActionsType }
